requirements.txt
absl-py==0.12.0
argon2-cffi==21.1.0
asgiref==3.4.1
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
clang==5.0
click==8.0.3
cloudpickle==2.0.0
cmake-setuptools==0.1.3
cudf @ file:///rapids/cudf-0%2Buntagged.1.g0a00579-cp38-cp38-linux_x86_64.whl
cugraph @ file:///rapids/cugraph-0%2Buntagged.1.gfa6f0f1-cp38-cp38-linux_x86_64.whl
cuml @ file:///rapids/cuml-0%2Buntagged.1.g273bc3c-cp38-cp38-linux_x86_64.whl
cupy-cuda114==9.3.0
cycler==0.11.0
Cython==0.29.24
dask==2021.7.1
dask-cuda==21.8.0
dask-cudf @ file:///rapids/dask_cudf-0%2Buntagged.1.g0a00579-py3-none-any.whl
dask-glm==0.2.0
dask-ml==1.9.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.4
distributed==2021.7.1
Django==3.2.6
entrypoints==0.3
fastavro==1.4.4
fastrlock==0.8
filelock==3.3.2
flatbuffers==1.12
fsspec==2021.7.0
future==0.18.2
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
graphsurgeon @ file:///workspace/TensorRT-8.0.3.4/graphsurgeon/graphsurgeon-0.4.5-py2.py3-none-any.whl
grpcio==1.39.0
gunicorn==20.1.0
h11==0.12.0
h5py==3.1.0
HeapDict==1.0.1
horovod @ file:///opt/tensorflow/horovod-source/dist/horovod-0.22.1.tar.gz
httptools==0.2.0
huggingface-hub==0.0.12
idna==3.3
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonschema==4.2.1
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyter-tensorboard @ git+https://github.com/cliffwoolley/jupyter_tensorboard.git@ffa7e26138b82549453306e06b535a9ac36db17a
jupyterlab==2.3.2
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
jupytext==1.13.1
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
llvmlite==0.36.0
locket==0.2.1
Markdown==3.3.4
markdown-it-py==1.1.0
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mdit-py-plugins==0.2.8
mistune==0.8.4
mock==3.0.5
msgpack==1.0.2
multipledispatch==0.6.0
nbclient==0.5.5
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
nltk==3.6.4
notebook==6.4.3
numba==0.53.1
numpy==1.19.4
nvidia-dali-cuda110==1.7.0
nvidia-dali-tf-plugin-cuda110==1.7.0
nvidia-dlprofviewer @ file:///opt/dlprof_viewer_install/nvidia_dlprofviewer-1.7.0-py3-none-any.whl
nvtx==0.2.3
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.2
pandas==1.2.5
pandocfilters==1.5.0
parso==0.8.2
partd==1.2.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==8.4.0
pluggy==1.0.0
polygraphy==0.33.0
portpicker==1.3.1
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.22
protobuf==3.17.3
psutil==5.7.0
ptyprocess==0.7.0
py==1.11.0
pyarrow @ file:///rapids/pyarrow-4.0.1-cp38-cp38-linux_x86_64.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.10.0
pynvml==11.0.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rmm @ file:///rapids/rmm-0%2Buntagged.1.gccd9647-cp38-cp38-linux_x86_64.whl
rsa==4.7.2
sacremoses==0.0.46
scikit-learn==0.23.1
scipy==1.4.1
Send2Trash==1.8.0
six==1.15.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tblib==1.7.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow @ file:///tmp/pip/tensorflow-2.6.0%2Bnv-cp38-cp38-linux_x86_64.whl
tensorflow-addons @ file:///opt/tensorflow/tf-addons/artifacts/tensorflow_addons-0.13.1-cp38-cp38-linux_x86_64.whl
tensorflow-datasets==3.2.1
tensorflow-estimator==2.6.0
tensorflow-metadata==1.4.0
tensorrt @ file:///workspace/TensorRT-8.0.3.4/python/tensorrt-8.0.3.4-cp38-none-linux_x86_64.whl
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
tftrt-model-converter==1.0.0
threadpoolctl==3.0.0
tokenizers==0.10.3
toml==0.10.2
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.9.1
treelite==1.3.0
treelite-runtime==1.3.0
typeguard==2.13.0
typing-extensions==3.7.4.3
ucx-py @ file:///rapids/ucx_py-0%2Buntagged.1.gbfa0450-cp38-cp38-linux_x86_64.whl
uff @ file:///workspace/TensorRT-8.0.3.4/uff/uff-0.6.9-py2.py3-none-any.whl
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.0
Werkzeug==2.0.2
whitenoise==5.3.0
wrapt==1.12.1
xgboost @ file:///rapids/xgboost-1.4.2-cp38-cp38-linux_x86_64.whl
zict==2.0.0
zipp==3.6.0
